Fix button overflow Move ToS
