+++ client/blocks/authentication/social/style.scss
@@ -108,6 +108,7 @@
 			letter-spacing: 0.32px;
 			margin-right: auto;
 			margin-left: -20px;
+			max-width: 250px;
 		}
 	}
 }
