+++ client/blocks/login/style.scss
@@ -208,6 +208,23 @@ $breakpoint-mobile: 782px; //Mobile size.
 	}
 }
 
+.login__form-subheader-terms {
+	max-width: 420px;
+	margin: 0 auto;
+
+	color: var(--studio-gray-50, #646970);
+	text-align: center;
+	font-size: $font-body-extra-small;
+	font-style: normal;
+	font-weight: 400;
+	line-height: 18px;
+
+	a {
+		color: var(--studio-gray-50, #646970);
+		text-decoration: underline;
+	}
+}
+
 .login__form-terms,
 .login__form-signup-link {
 	font-size: $font-body-extra-small;
